requirements.txt
python-dotenv==1.0.0
lxml==4.9.2
aiogram==2.25.1
bs4==0.0.1
fake-useragent==1.1.3
peewee==3.16.0
requests==2.28.2
psycopg2==2.9.5

